로컬 로직 마무리

+++ app/src/main/java/com/konkuk/kuit_kac/data/request/RequestDto.kt
@@ -27,5 +27,5 @@ data class FoodRequestDto(
     @SerialName("foodId")
     val foodId: Int,
     @SerialName("quantity")
-    val quantity: Int
+    val quantity: Float
 )
